Make the mapper as strict as the schemas actually are
A second cross-review pass probed the mapper directly instead of reading
its configuration, and the first pass's fix turned out not to deliver
what its own comment claimed. Six holes, each reproduced as a failing
test before being closed:

- a number or boolean was read into a string field as its text form;
- a number was read as an enum ordinal, so "stock_level": 1 became HIGH;
- a number or array was read as a date or timestamp, so
  "generated_at": 0 parsed as 1970-01-01;
- an explicit null in a required numeric field became 0, because
  primitives are defaulted before the constructor runs;
- a non-UTC offset was normalised to Z, changing the document on the way
  back out - every committed example uses Z, so nothing noticed;
- JSON numbers were narrowed to Double inside the parser, losing
  precision the schema allows and admitting NaN and infinities, which
  JSON cannot represent and Jackson emits as bare strings.

Money and temperature are now BigDecimal, which also gives the Market
Simulator an exact value to apply its scale-2 rounding to instead of an
already-rounded one. The two temporal types need a strict deserializer:
java-time reads a number as an epoch before coercion config is
consulted, so configuration alone cannot close that one.

design.md and spec.md previously claimed strictness the code did not
have. They now describe what it does, and name the two remaining
narrowings - integers only in canonical form, and Int for counts.

29 tests, up from 23.

+++ src/main/kotlin/club/podlodka/snowball/domain/PromotionOutcome.kt
@@ -2,19 +2,26 @@ package club.podlodka.snowball.domain
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.annotation.JsonPropertyOrder
+import java.math.BigDecimal
 
 /**
  * The `outcome` object of `promotion-outcome-v1.schema.json`.
  *
  * This is the whole business result the Market Simulator is allowed to publish: no coefficients,
  * no noise factor, no counterfactuals, no oracle-best action.
+ *
+ * `gross_profit` is a `BigDecimal` because the schema declares an unbounded JSON number and
+ * `docs/market-simulator/README.md` computes money at scale 2. `Double` would both lose precision
+ * the schema permits and admit `NaN` and infinities, which JSON has no way to represent. Carrying
+ * the exact value here does not decide the simulator's rounding policy; it only avoids destroying
+ * the result before the simulator can apply it.
  */
 @JsonPropertyOrder("units_sold", "gross_profit")
 data class PromotionOutcome(
     @JsonProperty("units_sold", required = true)
     val unitsSold: Int,
     @JsonProperty("gross_profit", required = true)
-    val grossProfit: Double,
+    val grossProfit: BigDecimal,
 ) {
     init {
         require(unitsSold >= 0) { "outcome.units_sold must not be negative, was $unitsSold" }
